backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
transformers==4.37.2
torch==2.1.2
torchaudio==2.1.2
numpy==1.26.3
pydantic==2.5.3
typing-extensions==4.9.0
accelerate==0.26.1
loguru==0.7.2
datasets==2.16.1
ffmpeg-python==0.2.0
soundfile==0.12.1